+++ src/briefy/leica/models/job/order.py
@@ -1,12 +1,10 @@
 """Briefy Leica Order to a Job."""
-from briefy.common.db.mixins import BriefyRoles
 from briefy.common.vocabularies.categories import CategoryChoices
 from briefy.leica.db import Base
 from briefy.leica.models import mixins
 from briefy.leica.models.job import workflows
 from briefy.leica.vocabularies import JobInputSource
 from briefy.ws.utils.user import add_user_info_to_state_history
-from briefy.ws.utils.user import get_public_user_info
 from sqlalchemy import orm
 from sqlalchemy.ext.hybrid import hybrid_property
 
@@ -22,7 +20,8 @@ __summary_attributes__ = [
 __listing_attributes__ = __summary_attributes__
 
 
-class JobOrder(mixins.OrderFinancialInfo, BriefyRoles, mixins.KLeicaVersionedMixin, Base):
+class JobOrder(mixins.OrderFinancialInfo, mixins.OrderBriefyRoles,
+               mixins.KLeicaVersionedMixin, Base):
     """A Job Order from the customer."""
 
     _workflow = workflows.JobOrderWorkflow
@@ -32,11 +31,6 @@ class JobOrder(mixins.OrderFinancialInfo, BriefyRoles, mixins.KLeicaVersionedMix
         ('view', ()),
         ('edit', ()),
         ('delete', ()),
-    )
-
-    __actors__ = (
-        'project_manager',
-        'scout_manager',
     )
 
     __colanderalchemy_config__ = {
@@ -262,23 +256,6 @@ class JobOrder(mixins.OrderFinancialInfo, BriefyRoles, mixins.KLeicaVersionedMix
         """
         project = self.project
         return project.tech_requirements
-
-    def _apply_actors_info(self, data: dict) -> dict:
-        """Apply actors information for a given data dictionary.
-
-        :param data: Data dictionary.
-        :return: Data dictionary.
-        """
-        actors = [(k, k) for k in self.__actors__]
-        info = self._actors_info()
-        for key, attr in actors:
-            try:
-                value = info.get(attr).pop()
-            except (AttributeError, IndexError):
-                data[key] = None
-            else:
-                data[key] = get_public_user_info(value) if value else None
-        return data
 
     def _summarize_relationships(self) -> dict:
         """Summarize relationship information.